Squish: Allow debugger test to run on Qt5
The test used to place a breakpoint in the line

    engine.load(QUrl(QStringLiteral("qrc:/qt/qml/qtquickapplication/main.qml")));

While Qt6 stops in that line once, Qt5 stops twice. Instead of handling
that difference just place the breakpoint in the line

    if (engine.rootObjects().isEmpty())

instead. This is always hit just once. Thanks to Karsten Heimrich for
the suggestion.

Currently, the lines are found by counting from the top of main.cpp
This shall be replaced by properly selecting the right line.

Change-Id: I0ab84bb1817d7dac22695c09539c14a5a43be11c
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>