Squish: Don't pass MSVS product line around when it's not needed
Since we're not testing MSVS2017 anymore, there are much less
differences between the product lines of MSVS. Most functions which
still got the product line as a parameter didn't use it anymore.
Instead, this now calls a function to get the product line only if,
when and where it's needed. The called function stores the
information so there is hardly any overhead but the code is cleaner.

Change-Id: Ib9b4f53e9f1b5fcd7e09a6ea7dddf487b4b6d8c2
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>