Squish: Work around invisible list items being inaccessible
That was not an issue earlier because the list of templates was
not long enough. The addition of the "Qt Test Application"
template then revealed the problem.

tst_new_project_defaults and -edit do not need this, because they
click one template after the other without closing the dialog.
That implicitly scrolls down the view.

Change-Id: Ie8edd4ba9ed91fa6852f75b88346cdb715300c2f
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>