Squish: Stabilize tst_debug on MSVS2019
When running the test on MSVS2019, Squish reliably switches to a
different application context. Depending on the timing, the test
may pass, fail or even crash. Eliminate this uncertainty by
stopping the test until the context change happened and setting
it back to the right context.
When running on MSVS2022, none of this seems to happen.

Change-Id: Ie1b1390afe274ae049aaf059014255249385fc63
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>