Squish: Wait for projects' files to be created
Usually, the files are already present when the test arrives there.
Only with CMake-based Qt Designer Custom Widget projects it may
take several seconds until they were written. MSVS seems to display
a progress bar during that time so this should not cause the test
to fail. Instead, wait up to a couple of seconds until the files
can be found.

Change-Id: Ic03855c098283b6dd6d778896c2d8cb4e4c3357f
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>