Let VS show a notification in case it cannot start the language client
We do not need to implement the ILanguageClient interface conditionaly,
depending on the version of the interface it will pick up the right
signature of OnServerInitializeFailedAsync(...). In case of VS 2019 we
now dump the exception to the output pane, VS 2022 will show a pop-up
in the left bottom corner with the failure message.

Change-Id: I965717619c3903713db6782f5465123fd7b09c9f
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>